% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_colors.R
\name{display_colors}
\alias{display_colors}
\title{Display the CSP color palette}
\usage{
display_colors(...)
}
\value{
A ggplot2 object
}
\description{
display_colors() produces a par plot showing the full color palette of the cspplot
package together with color names
}
