% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_colors.R
\name{list_colors}
\alias{list_colors}
\title{List the colors in the CSP color palette}
\usage{
list_colors()
}
\value{
a 2 x 14 tibble of the 14 colors in the CSP color palette and their Hex
color codes.
}
\description{
list_colors() returns a tibble of the CSP color names and their Hex color codes.
}
