% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_csp.R
\name{theme_csp}
\alias{theme_csp}
\title{ggplot2 theme for the AIDA course}
\usage{
theme_csp(
  title.size = 16,
  text.size = 14,
  legend.position = "top",
  show.axis = FALSE,
  show.grid = TRUE,
  plot.margin = c(0.2, 0.1, 0.2, 0.1)
)
}
\arguments{
\item{title.size}{title element size in pts}

\item{text.size}{text element size in pts}

\item{legend.position}{either one of "none", "left", "right", "bottom", "top"}

\item{show.axis}{boolean or "x", "y"; should axis be drawn? Which?}

\item{show.grid}{boolean; should grid lines be drawn?}

\item{plot.margin}{margin around entire plot (unit with the sizes of the top, right, bottom, and left margins)}
}
\description{
ggplot2 theme for the AIDA course
}
