% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_csp_dark.R
\name{theme_csp_dark}
\alias{theme_csp_dark}
\title{dark version of the CSP ggplot2 theme}
\usage{
theme_csp_dark(
  title.size = 16,
  text.size = 14,
  legend.position = "top",
  show.axis = FALSE,
  show.grid = TRUE,
  plot.margin = c(0.2, 0.1, 0.2, 0.1)
)
}
\arguments{
\item{title.size}{title element size in pts}

\item{text.size}{text element size in pts}

\item{legend.position}{either one of "none", "left", "right", "bottom", "top"}

\item{show.axis}{boolean or "x", "y"; should axis be drawn? Which?}

\item{show.grid}{boolean; should grid lines be drawn?}

\item{plot.margin}{margin around entire plot (unit with the sizes of the top, right, bottom, and left margins)}
}
\value{
ggplot theme
}
\description{
Theme for the "Cognitive Science of Pragmatics" Research Group, Department of Linguistics, University of Tübingen, Germany.
Based on the Aida theme (https://github.com/michael-franke/aida-package)
}
